추가 map 경로 생성 및 gps\ 공분산 시각화 완료

+++ src/Localization/localization/include/localization/GPS.hpp
@@ -42,9 +42,9 @@ class GPS{
 
         geometry_msgs::Point m_origin_coord;
         geometry_msgs::Point m_utm_coord;
-        float m_lat;
-        float m_lon;
-        float m_alt;
+        double m_lat;
+        double m_lon;
+        double m_alt;
         double m_utm_x;
         double m_utm_y;
         double m_utm_z;
